Add case_insensitive option.
Some people might want passwords to be treated case-insensitively.

As noted in the documentation, this needs to be treated with a little caution -
this option means that passwords are lower-cased before generating a hash, so if
you enable this, users with mixed-case passwords won't be able to log in any
more until their password is reset.  Similar applies if this option is turned
off.  So, if you want to use the option, decide at the start, and stick with it.